pyproject.toml
[tool.poetry]
name = "timetable-bot"
version = "0.1.0"
description = ""
authors = ["realaashil <singhalaashil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.3.0"
apscheduler = "^3.10.4"
datetime = "^5.4"
openpyxl = "^3.1.2"
pandas = "^2.1.4"
peewee = "^3.17.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"
requests = "^2.31.0"
paramiko = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
